Improve base adapters
